NOT WORKING: Refactor GUI initialization and solver selection
This commit refactors the GUI initialization in the main window and adds support for the Joule solver. The initial solver frame is now set correctly at startup. Additionally, the simulation controller now includes a Joule solver option. The simulation configuration has been updated to set the length manually to 1 when using the Fourier or Joule solver types. Finally, a new Joule frame has been added to the GUI for configuring Joule heating.